Cleaned up some code, fixed a leak of sorts.
When the code to activate the gui is triggered, it loops through all of
the players. Part of the check for opening the gui is that a player does
not already have a gui open, and inside that check it sets the gui being
open to true. This causes a dead code scenario, as all further checks in
the loop will be false. A break has been inserted to prevent being
wasteful of user resources. If this was a bug, and not an intended
effect, then it can be changed.